+++ pyinfraformat/core/__init__.py
@@ -1,7 +1,5 @@
 # pylint: disable=wildcard-import
 """Core pyinfraformat functionality."""
-
 from .core import *
 from .io import *
-from .parser import *
 from .utils import *
