+++ pyinfraformat/core/core.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from gc import collect
 import logging
 import os
+import numpy as np
 import pandas as pd
 
 from ..exceptions import FileExtensionMissingError
@@ -237,6 +238,7 @@ class Holes:
         else:
             df_list = [hole._get_dataframe(update=True) for hole in self.holes]
             self._dataframe = pd.concat(df_list, axis=0, sort=False)
+        collect()
         return self._dataframe
 
     def to_csv(self, path, **kwargs):
@@ -306,3 +308,183 @@ class Holes:
                 to_infraformat([hole], path_)
         else:
             to_infraformat(self.holes, path)
+
+
+class Hole:
+    """Class to hold Hole information."""
+
+    def __init__(self):
+        self.fileheader = FileHeader()
+        self.header = Header()
+        self.inline_comment = InlineComment()
+        self.survey = Survey()
+        self._illegal = Illegal()
+
+    def __str__(self):
+        from pprint import pformat
+
+        msg = pformat(self.header.__dict__)
+        return "Infraformat Hole -object:\n  {}".format(msg)
+
+    def __repr__(self):
+        return self.__str__()
+
+    def add_fileheader(self, key, fileheader):
+        """Add fileheader to object."""
+        self.fileheader.add(key, fileheader)
+
+    def add_header(self, key, header):
+        """Add header to object."""
+        self.header.add(key, header)
+
+    def add_inline_comment(self, key, comment):
+        """Add inline comment to object."""
+        self.inline_comment.add(key, comment)
+
+    def add_survey(self, survey):
+        """Add survey information to object."""
+        self.survey.add(survey)
+
+    def _add_illegal(self, illegal):
+        """Add illegal lines to object."""
+        self._illegal.add(illegal)
+
+    @property
+    def dataframe(self):
+        """Create pandas.DataFrame."""
+        return self._get_dataframe(update=False)
+
+    def _get_dataframe(self, update=False):
+        """Get pandas.DataFrame object.
+
+        Creates a new pandas.DataFrame if it doesn't exists of update is True
+
+        Paramaters
+        ----------
+        update : None or bool, optional, default None
+            None
+                Return earlier pandas.DataFrame if it exists and adds automatically new data
+            False
+                Return earlier pandas.DataFrame if it exists
+            True
+                Calculate a new pandas.DataFrame
+        """
+        if hasattr(self, "_dataframe") and not update:
+            if not self._dataframe.empty:
+                return self._dataframe  # pylint: disable=access-member-before-definition
+
+        dict_list = self.survey.data
+        if not dict_list:
+            logger.debug("No data in Hole object. Header: {}".format(self.__str__()))
+            self._dataframe = pd.DataFrame()
+            return self._dataframe  # pylint: disable=attribute-defined-outside-init
+        self._dataframe = pd.DataFrame(dict_list)  # pylint: disable=attribute-defined-outside-init
+        self._dataframe.columns = ["data_{}".format(col) for col in self._dataframe.columns]
+        if not self._dataframe.empty:
+            self._dataframe.loc[0, self._dataframe.columns] = np.nan
+        for key in self.header.keys:
+            self._dataframe.loc[:, "Date"] = self.header.date
+            for key_, item in getattr(self.header, key).items():
+                self._dataframe.loc[:, "header_{}_{}".format(key, key_)] = item
+        for key in self.fileheader.keys:
+            for key_, item in getattr(self.fileheader, key).items():
+                self._dataframe.loc[:, "fileheader_{}_{}".format(key, key_)] = item
+
+        return self._dataframe
+
+
+class FileHeader:
+    """Class to hold file header information."""
+
+    def __init__(self):
+        self.keys = set()
+
+    def add(self, key, values):
+        """Add member to class."""
+        setattr(self, key, values)
+        self.keys.add(key)
+
+    def __str__(self):
+        msg = "FileHeader object - Fileheader contains {} items".format(len(self.keys))
+        return msg
+
+    def __repr__(self):
+        return self.__str__()
+
+    def __getitem__(self, attr):
+        if attr in self.keys:
+            return getattr(self, attr)
+        else:
+            raise TypeError("Attribute not found: {}".format(attr))
+
+
+class Header:
+    """Class to hold header information."""
+
+    def __init__(self):
+        self.keys = set()
+        self.date = pd.NaT
+
+    def add(self, key, values):
+        """Add header items to object."""
+        if key == "XY" and ("Date" in values):
+            if len(values["Date"]) == 6:
+                date = datetime.strptime(values["Date"], "%d%m%y")
+            elif len(values["Date"]) == 8:
+                date = datetime.strptime(values["Date"], "%d%m%Y")
+            else:
+                try:
+                    date = pd.to_datetime(values["Date"])
+                except ValueError:
+                    date = pd.NaT
+            self.date = date
+        setattr(self, key, values)
+        self.keys.add(key)
+
+    def __getitem__(self, attr):
+        if attr in self.keys:
+            return getattr(self, attr)
+        else:
+            raise TypeError("Attribute not found: {}".format(attr))
+
+
+class InlineComment:
+    """Class to inline comments."""
+
+    def __init__(self):
+        self.data = list()
+
+    def add(self, key, values):
+        """Add inline comments to object."""
+        self.data.append((key, values))
+
+    def __getitem__(self, attr):
+        return self.data[attr]
+
+
+class Survey:
+    """Class to survey information."""
+
+    def __init__(self):
+        self.data = list()
+
+    def add(self, values):
+        """Add survey information to object."""
+        self.data.append(values)
+
+    def __getitem__(self, attr):
+        return self.data[attr]
+
+
+class Illegal:
+    """Class to contain illegal lines."""
+
+    def __init__(self):
+        self.data = list()
+
+    def add(self, values):
+        """Add illegal lines to object."""
+        self.data.append(values)
+
+    def __getitem__(self, attr):
+        return self.data[attr]
